Add files via upload
Serial monitor transmits data from Arduino IDE (C) to python. Please note: Python code shouldn't be run while serial monitor window is open. In this lab a clean sine wave has noise added to it on the arduino and C side. Next Python filters the signal and finally the sampling rate shows the ability to reproduce a clean sine wave as well as the accuracy or squareification of the wave depending on how high/low the nyquist sampling rate is. Finally, the output of the nyquist frequencies is plotted.